address PR feedback: rename views and flatten ProposalSection
